Simple Rig

+++ common/include/ecs/comp_renderer.h
@@ -24,8 +24,8 @@ typedef struct comp_renderer_constructor_info {
   phy_rigid_body *body; 
   phy_shape *shape; 
   const int tex_id;
-  const size_t rig_id;
-  const size_t area_id;
+  const unsigned int rig_id;
+  const unsigned int area_id;
   const int render_layer;
   const float parallax_value;
   const bool flipped;
