+++ common/src/ecs/comp_animator.c
@@ -16,7 +16,7 @@ static ecs_ret sys_animator_update(ecs_ecs *ecs, ecs_id *entities, const unsigne
   if (dt == 0.0f) {
     return 0;
   }
-  for (size_t id = 0; id < entity_count; id++) {
+  for (unsigned int id = 0; id < entity_count; id++) {
     comp_animator *info = ecs_get(ECS, entities[id], ANIMATOR_COMPONENT_SIGNATURE);
     comp_renderer *r_info = ecs_get(ECS, entities[id], RENDERER_COMPONENT_SIGNATURE);
     if (!r_info->rig.enabled) {
